# src/api-spec/ontology.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tellus Ontology Engine \u2014 Full API",
    "description": "Comprehensive API specification for the Tellus Ontology Engine. Covers ontology CRUD, object types, properties, link types, and action types.",
    "version": "2.0.0",
    "contact": {
      "name": "Tellus Engineering",
      "email": "engineering@tellus.dev"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v2/usage": {
      "post": {
        "tags": [
          "Usage"
        ],
        "summary": "Log a usage event",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "resourceType",
                  "operation"
                ],
                "properties": {
                  "ontologyId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "resourceType": {
                    "type": "string"
                  },
                  "resourceId": {
                    "type": "string"
                  },
                  "userId": {
                    "type": "string"
                  },
                  "operation": {
                    "type": "string"
                  },
                  "appId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usage event logged"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/usage": {
      "get": {
        "tags": [
          "Usage"
        ],
        "summary": "Query usage events",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "resource",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage events"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/proposals": {
      "get": {
        "tags": [
          "Proposals"
        ],
        "summary": "List proposals",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of proposals"
          }
        }
      },
      "post": {
        "tags": [
          "Proposals"
        ],
        "summary": "Create a proposal",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title"
                ],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "branchId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "author": {
                    "type": "string"
                  },
                  "reviewer": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Proposal created"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/proposals/{proposalId}": {
      "put": {
        "tags": [
          "Proposals"
        ],
        "summary": "Update a proposal",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "proposalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Proposal updated"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/security-policies": {
      "get": {
        "tags": [
          "Security Policies"
        ],
        "summary": "List security policies",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of policies"
          }
        }
      },
      "post": {
        "tags": [
          "Security Policies"
        ],
        "summary": "Create a security policy",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "policyType"
                ],
                "properties": {
                  "objectTypeApiName": {
                    "type": "string"
                  },
                  "policyType": {
                    "type": "string",
                    "enum": [
                      "row_level",
                      "column_level",
                      "role"
                    ]
                  },
                  "config": {
                    "type": "object"
                  },
                  "active": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Policy created"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/security-policies/{policyId}": {
      "put": {
        "tags": [
          "Security Policies"
        ],
        "summary": "Update a security policy",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Policy updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Security Policies"
        ],
        "summary": "Delete a security policy",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "policyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          }
        }
      }
    },
...
  },
...
}